refactor base renderer
renew package-lock.json
